Seller can change the status of the order to  SHIPPED.

+++ E-commerce/src/main/java/com/example/E_commerce/controller/ProductController.java
@@ -10,10 +10,13 @@ import com.example.E_commerce.exception.ResourceNotFoundException;
 import com.example.E_commerce.mapper.ProductMapper;
 import com.example.E_commerce.service.ProductService;
 import com.example.E_commerce.service.UserService;
+import com.example.E_commerce.service.security.service.JWTservice;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
@@ -26,8 +29,10 @@ public class ProductController {
     private UserService userService;
     @Autowired
     private ProductService productService;
-    @PostMapping("/add/{sellerId}")
-    public ResponseEntity<?> addProduct(@Valid @RequestBody ProductRequestDTO productRequestDTO, @PathVariable int sellerId){
+    @PreAuthorize("hasRole('SELLER')")
+    @PostMapping("/add")
+    public ResponseEntity<?> addProduct(@Valid @RequestBody ProductRequestDTO productRequestDTO){
+        int sellerId= JWTservice.getAuthenticatiedUser().getId();
         Product product= new Product();
         product.setName(productRequestDTO.getName());
         product.setPrice(productRequestDTO.getPrice());
@@ -43,10 +48,16 @@ public class ProductController {
         productService.save(product);
         return ResponseEntity.ok(ProductMapper.createProductDTO(product));
     }
+
+    @PreAuthorize("hasRole('SELLER')")
     @PutMapping("/updateproduct/{productId}")
     public ResponseEntity<ProductDTO> updateProduct(@Valid @RequestBody ProductRequestDTO productRequestDTO, @PathVariable int productId){
+        int sellerId= JWTservice.getAuthenticatiedUser().getId();
         Product product= productService.findById(productId).orElseThrow(()->
                 new ResourceNotFoundException("Cannot Find the product with Specified Id"));
+        if(product.getSeller().getId()!=sellerId){
+            throw new AccessDeniedException("Sorry, the Product with provided Id donot belongs to you");
+        }
         product.setName(productRequestDTO.getName());
         product.setPrice(productRequestDTO.getPrice());
         product.setDescription(productRequestDTO.getDescription());
@@ -58,17 +69,29 @@ public class ProductController {
         productService.save(product);
         return ResponseEntity.ok(ProductMapper.createProductDTO(product));
     }
+
+    @PreAuthorize("hasRole('SELLER')")
     @DeleteMapping("/delete/{productId}")
     public ResponseEntity<String> deleteProduct(@PathVariable int productId){
+        int sellerId= JWTservice.getAuthenticatiedUser().getId();
+        Product product= productService.findById(productId).orElseThrow(()->
+                new ResourceNotFoundException("Cannot Find the product with Specified Id"));
+        if(product.getSeller().getId()!=sellerId){
+            throw new AccessDeniedException("Sorry, the Product with provided Id Do not belongs to you");
+        }
         productService.delete(productId);
         return ResponseEntity.ok("SuccessFully Deleted");
     }
+    @PreAuthorize("permitAll()")
     @GetMapping("/findbyid/{productId}")
     public ResponseEntity<ProductDTO> findById(@PathVariable int productId){
+        int sellerId= JWTservice.getAuthenticatiedUser().getId();
         Product product= productService.findById(productId).orElseThrow(()->
                 new ResourceNotFoundException("Cannot find the product with provided id"));
         return ResponseEntity.ok(ProductMapper.createProductDTO(product));
     }
+
+    @PreAuthorize("permitAll()")
     @GetMapping("/findall")
     public ResponseEntity<List<ProductDTO>> findALl(){
         List<Product> productList= new ArrayList<>();
